Update FinalCode
Changed name of space from “2D space” to “Pond for Project”.

Used Class to define my object which in this case is a moving object that is also the fish. 

Used the init__() function as the constructor for the Class. So when a new instance of "MovingObject" occurs, this function will initialize the attributes listed. These are X, Y, width, height, color, name. 

Self is the first parameter and thus is the object in question.
Self.X and Self.Y represent coordinates of the object.
Self.Width/Height/Color/Name represent their physical attributes..
Self.velocity and self.direction were used to give a random speed and direction to each object to mimic a realistic environment of fish. I am trying to figure out a way for the fish to move more fluidly and not just bounce off the walls but have not figured out a way to do that yet.
*********************
Def draw just paints the image of the fish onto the screen. In this case, I used pygame.draw.rect and a rectangle will be used to represent the fish. The attributes such as color, name, etc. are attached to each rectangle.
I also assigned the name of each fish to itself and that is seen using font and text. Text rect just centers the name.
*********************
Def move just makes it so the fish do not leave the boundaries of the screen, as to mimic an enclosed, limited space.