Update FinalCode
#Main Loop is the loop that keeps the fish moving on screen. It enables the program to function.

Not sure if screen.fill is necessary, but from what I understand, it clears any previous instances of the fish being in a certain space. So really, it just assists in what looks like the movement of the fish itself as it clears the screen briefly to show the next position of the fish.

FIsh move and draw just do what they say. Present the visual.